Change documentation for intersection method of abstract properties and fix test suites.
Intersection was declared to return a lower bound. Actually, this is not consistent with the
rest of the operation which are upper approximations of some concrete operations. Therefore,
we now define intersection to be an upper approximation of the concrete glb operator.

Mostly, this will be exactly the glb, and in this case a new trait for the test suite has
been added.